+++ tybotRoute/tiledeskChatbotPlugs/directives/Directives.js
@@ -21,6 +21,7 @@ class Directives {
   static IF_ONLINE_AGENTS = "ifonlineagents";
   static FUNCTION_VALUE = "functionvalue";
   static JSON_CONDITION = "jsoncondition";
+  static JSON_CONDITION_2 = "jsoncondition2"; // NEW: azione JSON Condition V2 del Design Studio
   static SET_ATTRIBUTE = "setattribute";
   static SET_ATTRIBUTE_V2 = "setattribute-v2";
   static REPLY = 'reply';